+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `CardControllerTest`
 - Added `DashboardControllerTest`
 - Added `ResourceControllerTest`
+- Adopted configuration
 
 ## [2.4.9] - 2026-01-17
 - Fetch XSRF-Token early